cmd/snap-update-ns: add opt-foo test case data
This test case uses a layout on /opt/foo to expose content from another snap.
The refresh test case is more interesting, as it shows how the content is
refreshed to a new revision, and that new revision is actually visible in
practice.

The data set is not yet used in practice, but shows a new way of collecting
data that is less repetetive, and how the tested release of snapd is faulty.

Test data was generated with snapd 2.66

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>