multiple: revert adding SplitMountOptions
SplitMountOptions was originally introduced to account for commas
inside paths that are passed as options to systemd-mount. i.e for the
case of overlayfs mounts, the lowerdir/upperdir/workdir options could
contain paths with commas (which are valid path names).

For now we are going to restrict backslashes, colons, double quotes,
spaces and commas from these paths therefore it is not needed.