cmd/snap-update-ns: add opt test case data
This test case uses a layout on /opt to expose content from another snap. The
refresh test case is more interesting, as it shows how the content is refreshed
to a new revision, and that new revision is actually visible in practice.

Out of all the new test cases, this is the only one that reliably works with
know patches, not yet merged into snapd. All the more complicated versions
exhibit some form of failure.

The data set is not yet used in practice, but shows a new way of collecting
data that is less repetitive, and how the tested release of snapd is faulty.

Test data was generated with snapd 2.66
                                                                                                                                                    The data set is not yet used in practice, but shows a new way of collecting
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>