c/snap-bootstrap: unquote options to systemd-mount
double quoting the options to systemd-mount resulted in errors for the
overlayfs case such as:

localhost kernel: overlayfs: failed to resolve '"/run/mnt/foo"': -2

which is possibly due to an older util-linux being used.

In the same spirit, since we now forbid many characters, we can safely
remove the extra double quotes.