run-checks: run golangci-lint on all files
Running golangci-lint with --new-from-rev will only run it against the
diff which is not enough to determine if the change adds issues (e.g.,
if it removes the last call to a function). Another alternative would be
to still run it with --new-from-rev but also --whole-files which would
mitigate it but not completely eliminate the chance of this happening.

Signed-off-by: Miguel Pires <miguelpires94@gmail.com>