cmd/snap-update-ns: remove incorrect optimization
Remove the complex logic to compute a delta to be applied by snap-update-ns and
replace it with a unmount-all, mount-all approach. A number of tests are
removed as they were testing some of the previous elaborate behavior. I saw no
need to keep them in absence of said behavior. Some tests are skipped and need
to be evaluated further. One integration test that used to fail now passes.

Since this is a rather big change, it warrants discussion of some of the
history as well as some of the alternative approaches and the path forward.

** The problem **

Broadly speaking, computing a diff of two fstab-like mount profiles with the
goal to know what to mount and what to unmount only works if we are mounting
real file systems that are using unique device names as identifiers. It does
*not work*, when the mount profile represents bind mounts, as then the source
path is only a symbol and not a sufficient identifier of the source material.
The actual source material depends on all prior mounts and the final shape of
the mount namespace. Our long path towards realising this is littered with
complex behavior of the optimizer pass, and bugs affecting applications.

In response to bugs in this area snapd has evolved a more and more elaborate
algorithm that tries to tailor to specific cases in one way or another. I
strongly believe all that logic is wrong. Instead of special cases, we should
have not done the optimization or really come up with a full model of mount
trees, bind mounting and propagation. At some point though, the cost exceeds
any perceived benefit, in my opinion.

The recently added test files, present in cmd/snap-update-ns/testdata/ show how
badly the current code fares in face of relatively simple operations performed
on snaps that follow a common idiom present in nearly all snap packages with
graphical applications.

I did my best to look for bugs that relate to this change. I'm sure I've missed
some. The list I managed to complete is at the end of the commit message.

** Origin of persisted mount namespaces **

In the original 2.x release, snapd has shipped with a system of persistent
mount namespace associated with each snap instance. The motivation back then
was two-fold: on one hand side this allowed the per-snap private /tmp directory
to keep its state from one invocation of a given snap to another. On the other
hand side it allows avoiding the cost of mount namespace setup to be paid on
each invocation. Instead, the first invocation of a given snap instance,
subject to some limitations, sets up the mount namespace. First by running
snap-confine to do the initial, mostly unconditional preparation which includes
pivot-root.  Then followed by a call to snap-update-ns in inherited-namespace
mode, to allow it to set up additional mount entries as dictated by the mount
profile of the snap, lastly followed by an optional setup of per-snap, per-user
mount namespace again unshared from the per-snap mount namespace - conditional
on the presence of the per-user mount profile.

All of this leaves us with at between two and three mount namespaces. The mount
namespace of the init process, the mount namespace of the given snap,
initialled and persisted at /run/snapd/ns/$SNAP_INSTANCE_NAME.mnt and an
optional per-user mount namespace which is not persisted by default, although
this is supported with a conditional feature that has seen remarkably little
testing in face of the history of bugs in this area.

** Updating in place **

Historically the approach to updating the persisted state was based on the
assumption that we can look at two fstab-like mount profiles and compute a
delta. The removed lines would get unmounted, added lines would get mounted and
unchanged lines would stay as-is. This approach does work for the simples use
cases, ideally those that show a content being mounted at an existing directory
in $SNAP or even a dynamically created directory in $SNAP_COMMON or $SNAP_DATA.
The list of cases where this behaves correctly is sadly remarkably short and
without going into needless details, one can assume, ends at the examples just
given.

Real world application authors have come up with a number of cases that behave
less than ideally under typical conditions. There are dozens of bugs on
Launchpad and elsewhere, describing mysterious failure modes. One can
characterize those into two broad groups: one where content is unexpectedly
missing and, a more subtle one, where content is unexpectedly stale. Since the
second case is usually hard to notice and rarely results in application
crashes, one can assume that the bulk of the reported failures are a result of
the first-case. A connected content, typically holding shared libraries, is not
available despite the relevant snap interface being connected, the mount
profile looking just right.

** Alternatives **

During the work on this problem I've considered some alternatives. Those include:

- Stop persisting mount namespaces entirely. This would mostly work except for
  some important snaps that live on the edge of what normal snaps are allowed
  to do. The primary example is LXD and the way it uses the mount namespace to
  store further mount namespaces.  Not persisting anything would just break it
  entirely, or force a re-design.

- Attempting to model kernel mount namespaces, with perhaps, bug-for-bug
  behavior in order to model what would happen to understand if an optimization
  can be taken. I've deemed too complex to be practical.

- Alternative sorting schemes. I think this one still has merit but not in
  isolation.  As a follow-up to the mount story I would like to sort the
  desired mount profile with an algorithm that understands bind mounts better.
  I strongly believe building a graph or lattice and particular traversal of
  that graph would provide a generic solution to special-cases of how
  layouts/content interact. With this we might be able to entirely remove
  special cases that look at the type of mount entry.

Fixes: https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/2043993
Fixes: https://bugs.launchpad.net/snapd/+bug/1828354
Fixes: https://bugs.launchpad.net/snapd/+bug/1828357
Fixes: https://bugs.launchpad.net/snapd/+bug/2034056
Fixes: https://bugs.launchpad.net/snapd/+bug/1856093
Fixes: https://bugs.launchpad.net/snapd/+bug/2055273
Jira: https://warthogs.atlassian.net/browse/SNAPDENG-31645

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>