cmd/snap-update-ns: switch to a three-pass process
Historically Change.Perform was a two-step process, first it would ensure that
both the source and destination of the operation are ready and only then it
would call the lowLevelPerform function to actually mount or unmount.

This works for simple cases but has the problem that the desired changes are
interleaved with construction of writable mimics, especially in
ensure{Source,Target}. This is a problem because it may end up performing a
sequence of operations where a mount we perform is duplicated by the mimic,
like in the example given below:

High-level view:
- unmount /target
- mount --bind /source1 /target/dir
- mount --bind /source2 /target/dir/subdir

Low-level view:
- umount --lazy /target
- make writable mimic /target/dir
- mount --bind /source1/ /target/dir
- make writable mimic /target/dir/subdir
- mount --bind /source2 /target/dir/subdir

The bind mount /target/dir now has two instances, one performed directly and
then another that was created during the construction of a mimic at
/target/dir/ in order to create the sub-directory subdir.

The new three-pass approach breaks that to:
 - apply all keep and unmount actions
 - prepare to apply all mount actions by ensuring source/target is ready and
   constructing any writable mimics necessary.
 - actually apply all mount actions

The new order of operations, for the example given above, is now:

New low-level view:
- umount --lazy /target
- make writable mimic /target/dir
- make writable mimic /target/dir/subdir
- mount --bind /source1/ /target/dir
- mount --bind /source2 /target/dir/subdir

The functions related to preparation are now called PrepareToPerform and the
functions related to actually doing the change are called DoPerform. A helper
that calls one and then the other remains so that fewer overall changes are
necessary. The lowLevelPerform method is now gone and is now the actual
implementation of DoPerform.

Test mocking functions can now mock both sides individually. Mocking was also
redesigned from the classical function variable due to the cycle caused by
PrepareToPerform calling DoPerform internally.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-23293

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>