tests: run nested tests using ubuntu noble as host
This change change the image used as ubuntu host in nested tests.

So far we used xenial for uc16, bionic for uc18, etc, which is
unconvenient because we need to maintain several images and the code
complexity also is higher.

The system names went from ubuntu-xx.04-64 to ubuntu-xx-64, this is
because we run ubuntu core and classic systems in the nested vms. Also
it is convenient because otherwise we should change for example all the
systems in the tests from ubuntu-2* to ubuntu-core-2* (or similar).

Also the dependency packages installed in the neste host was reduced and
the spread.yaml was simplified.