i/builtin: escape @ in custom-device filepaths and prohibit in devices
Since `@` is a special character in AppArmor used to denote variables,
all occurrences of literal `@` in filepaths must be escaped with `\` in
AppArmor rules. If a `@` is already escaped, do not add another `\`,
else the `\` will escape the existing `\` and leave the `@` unescaped.

This escaping should occur when the AppArmor spec is generated, rather
than during validation, so the presence of `@` characters is not checked
during validation.

Also, as `@` should not occur in device paths (under `/dev/`), add `@`
to the disallowed characters regex to preserve the current behavior.
This may be reconsidered in the future if users report needing `@` in
device paths instead of just in read/write filepaths.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>