i/p/requestrules,o/i/apparmorprompting: allow overlapping rules (#14538)
* i/p/requestrules,o/i/apparmorprompting: allow overlapping rules

Allow several rules which render to the same variant to coexist in the
tree without conflict, so long as the outcome of all those overlapping
rules is identical.

This allows the client to reply with "allow read forever for /foo/bar"
and then later say "allow read|write forever for /foo/bar" without the
latter being treated as a rule conflict error. Clearly, the second rule
is a superset of the first, and there's no intent-based reason that
these two rules couldn't coexist, it was just an implementation detail
that we previously only allowed a pattern variant to be associated with
a single rule ID.

Now, each pattern variant in the tree for a particular snap, interface,
and permission can be associated with a set of rule IDs. Any non-expired
rules in that set must have the same outcome. Any expired rules in the
set are ignored (and removed when convenient).

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestrules: add clarifying comment about match outcome

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestrules: associate outcome with variant entry and clarify logic

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestrules: simplify closure which adds rule to tree

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>