c/snap-bootstrap/initramfs_systemd_mount: Add support for overlayfs options and allow empty what argument for tmpfs (#14591)
* c/snap-bootstrap/initramfs_systemd_mount: add support for overlayfs mount options

* c/snap-bootstrap/initramfs_systemd_mount: allow empty what argument in doSystemdMountImpl

* c/snap-bootstrap: make check for overlayfs arguments more explicit

* c/snap-bootstrap: escape paths for overlay mounts

and modify the testsuite to account for escaped commas that could
exist in the options argument as part of the overlay paths.

* c/snap-bootstrap: add support for multiple overlayfs lowerdirs

* c/snap-bootstrap: throw error for insufficient arguments for overlayfs

* c/snap-bootstrap: fix test case

* c/snap-bootstrap: remove stray debug command from test

* c/snap-bootstrap: revert to single escaping for overlay options

upon closer inspection, double escaping is not needed when
invoking systemd-mount through Go's exec as it doesn't spawn
a shell.

This was smoked tested using an unmocked version of
TestDoSystemdMount and by performing a mount with a lower directory
that includes a comma.

* multiple: move split mount options logic to osutil

fstab entries with overlay options that could contain commas in the
path could not potentially be parsed properly. The logic that was
implementing that functionality was moved from a test case in
snap-bootstrap to mountentry_linux.go in osutil.

It was also identified that the moved code was causing one of the
tests to fail (TestParseMountEntry3). This bug was fixed.

A new test case was also added.

* multiple: clarify splitMountOptions intended usage

* c/snap-bootstrap: escape spaces in systemd mount options

* osutil/mountentry_linux_test.go: add tests for SplitSystemdMountOptions

* osutil/mountentry_linux.go: revert to old behaviour for ParseMountEntry

* c/snap-bootstrap: add clarifying comment for multiple overlayfs lower layers

* osutil/mountentry_linux.go: merge splitMountOptions into SplitSystemdMountOptions

* osutil/mountentry_linux.go: add jira id for TODO item

* c/snap-bootstrap: use a separate replacer for overlayfs lowerdirs

* c/snap-bootstrap: disallow backslash in all systemd-mount paths

* c/snap-bootstrap: disallow backslashes, commas, spaces and colons

from appearing in paths used in the systemd-mount wrapper. The exception
of this is paths used in the lowerdir option of overlayfs mounts.

* c/snap-bootstrap: disallow double quotes too

from appearing in paths used in the systemd-mount wrapper.

* c/snap-bootstrap: remove restrictions for what and where arguments

* multiple: revert adding SplitMountOptions

SplitMountOptions was originally introduced to account for commas
inside paths that are passed as options to systemd-mount. i.e for the
case of overlayfs mounts, the lowerdir/upperdir/workdir options could
contain paths with commas (which are valid path names).

For now we are going to restrict backslashes, colons, double quotes,
spaces and commas from these paths therefore it is not needed.

* fixup! c/snap-bootstrap: remove restrictions for what and where arguments

* c/snap-bootstrap: refactor pathContainsAny

* c/snap-bootstrap: unquote options to systemd-mount

double quoting the options to systemd-mount result in errors such as

localhost kernel: overlayfs: failed to resolve '"/run/mnt/foo"': -2

which is possibly due to an older util-linux being used.

Since we now forbid many characters, we can safely remove the extra
double quotes.

* c/snap-bootstrap: specify forbiddenChars for overlayfs lowerdir variable name

* c/snap-bootstrap: forbid colons from overlayfs lowerdir paths too

- fix appending of multiple lowerdirs.
- add a test for a lowerdir with a colon.

* c/snap-bootstrap: address minor comment