timeutil: fix timer stop/reset return values and add more methods
Rather than returning whether the timer had fired, `Stop` and `Reset`
should return whether the timer is currently active. If it has fired or
been stopped manually, these should return false.

Additionally, add the `Active` method to report whether the timer is
currently active, and add the `FireCount` method to report the number of
times that the timer has fired. This is useful to directly check whether
a timer's callback has been called.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>