interfaces/apparmor: don't grant access to systemd userdb API sockets
These sockets provide access to the varlink-based API of systemd-userdbd which
in turn exposes hashed passwords and other details of the current user account
to any snap running as that user, as it doesn't distinguish between sandboxed
applications compared to traditional applications running as the user. In turn,
if the snap runs a root daemon it can then access the account details (again
including hashed passwords) of any user on the system.

Note, the impact of this issue is limited since systemd-userdbd is not installed
by default in Ubuntu or other distributions that I am aware of and so in this
case these varlink sockets are not active and no information is exposed.

This commit fixes LP: #2067006 by removing this access entirely - note this
issue was assigned CVE-2024-5300.

Signed-off-by: Alex Murray <alex.murray@canonical.com>