o/ifacestate: fix generation of AppArmor profile w/ incorrect revision
In a scenario where two snaps with connected content plugs and slots
were refreshed together, the first snap to be refreshed could get an
AppArmor profile generated that used the previous revision (so
revisioned paths like $SNAP_DATA and $SNAP_USER_DATA) would not be
accessible. The issue is that we generate security profiles for
connected snaps when generating the profile for the refreshed snap but
don't take into account that the snap may have been refreshed in the
same change (the new revision isn't reflected in state yet). So we'd
generate an incorrect AppArmor profile with the previous revision.
I couldn't add a spread test to tests/regression/ because it's not a
deterministic failure.

Signed-off-by: Miguel Pires <miguelpires94@gmail.com>