testtime: replace internal expiration/currtime with duration/elapsed
Since `TestTimer` is designed to allow the passage of time to be
controlled manually, it makes sense to represent this internally using a
duration and elapsed time, rather than storing the timestamp at the
creation of the timer and advancing an internal "current time".

This way, the time at which the timer was created has no effect on the
expiration timestamp which may be sent over the C channel when it fires,
and timers don't each have their own conflicting view of what time it
is. When a timer expires, the current time is sent over the channel, or
if `Fire` is called directly, then the given time is sent instead.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>