i/builtin: prohibit trailing '@' in common-files filepaths (#14696)
* i/builtin: prohibit trailing '@' in common-files filepaths

For the `common-files` interface and those which use
`commonFilesInterface` (such as `system-files` and `personal-files`),
these append a trailing `{,/,/**}` to the end of each path.

The `@` character is allowed in filepaths specified in these plugs,
since filepaths like `/dev/foo@bar` may be required. However, if an `@`
character occurs at the end of the filepath (e.g. `/foo/bar@`), then the
resulting filepath in the rule is `/foo/bar@{,/,/**}`. AppArmor treats
`@{FOO}` as a variable, so having `@{,/,/**}` in the final rule snippet
is problematic.

This commit adds a check to `common_files.go` to ensure that filepaths
do not have a trailing `@`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/builtin: add comment explaining why common-files paths cannot end with @

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>