tests: fix raciness in async startup and sync install
The test installs a snap, starts it in the background and proceeds to
pack and install a test snap that would affect the mount namespace of
the racing-to-start program.

In my local testing I can easily hit the case that the mount namespace
is not yet constructed by the time the second snap is installed,
somewhat defeating the point of the test.

Synchronously construct the mount namespace before starting the async
loop to prevent this from happening.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>