tests: fix broken app definition
The app "cmd" had an incorrect meta/snap.yaml entry - cmd instead of
command. I suspect this is why the whole test ran under snap run --shell
but it is hard to know that for sure.

Introduce a "standard" sh app that runs and behaves as /bin/sh and
adjust the test to match.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>