tests/lib/tools/tests.nested: change --tpm-no-restart to --keep-firmware-state
When restarting a nested VM, we reset the EFI variables and the TPM.
It was possible to keep the state of the TPM, but it makes little
sense to keep the TPM but not the EFI variables. With the new secboot,
both resetting EFI and resetting the TPM will potentially cause
recovery key to be needed.

This fixes the case where an installer image would create a sbatlevel
revocation, and get a different one after reset and boot with the
gadget snap.