o/{ifacestate,snapstate}: delay reboot when we have kernel-modules
Delay reboot when we are installing/updating/removing kernel-modules
components jointly with a kernel refresh so it happens after the
refresh hooks have run. With this we ensure that the kernel modules
shipped in the component or built from the hooks is available at the
same time as the other modules after rebooting.