secboot: register keys in keyring with also the legacy paths (#14718)
snap-bootstrap use to unlock encrypted disks using
/dev/disk/by-partuuid paths.  This was used ask identifier for the
keys stored in the keyring. However /dev/disk/by-uuid path makes more
sense as UUID is a part of LUKS2, and cryptsetup uses it for the
identify of the volume key. Partition UUID on the other hand are only for
GPT partitions.

However older snapd will query the key using the old path. So we need
to make sure snap-bootstrap also register the old name. A new
snap-bootstrap should always be able to run and older snapd.