o/snapstate: always run validate-component when installing a component from the store
Even if the component happens to already be installed with a previous
snap revision, we will need to download a new snap-resource-pair
assertion, which is done by validate-component.

Note that this doesn't run validate-component when installing asserted
components from a local snap file. However, sideloading components is
not yet supported by the public snapd API, so that can be handled later.