s/i/dmverity/veritysetup: add superblock parsing functionality
adding helpers for retrieving and parsing a dm-verity superblock from a
dm-verity hash device/file. This will be first consumed by the snap
integrity API which will need to detect the salt that was used for the
dm-verity data generation. Moreover callers to dmverity.Format will
need to have a way to retrieve the parameters used by veritysetup if
no parameters are passed (and veritysetup chooses default values).