i/p/requestprompts: remove racy timer stop during close and clarify behavior
The `Close` method calling `Stop` on the expiration timers races with
those timers firing, and it's unnecessary since the the callback
function checks whether the DB has been closed and returns early if it
has. Thus, remove the explicit `Stop` and add some comments about the
reasoning behind this.

Additionally, add a `isClosed` helper so functions don't have to
directly call a method on the contained `maxIDMmap` to check whether the
prompt DB has been closed.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>