t/n/m/uc-update-assets-secure: sign new grub also with old key [backport] (#14288)
* t/n/m/uc-update-assets-secure: sign new grub also with old key

It can happen that an update of assets in seed partition is interupted
and shim is old, but grub is new. That means that the new grub has to
be signed with a vendor key from shim.

Secboot will refuse to reseal if we have a potential impossible boot
chain. So we should not test that impossible case.

* tests/nested/manual/uc-update-assets-secure: fix signing of fallback binary