i/builtin: add the checkbox-support interface
The checkbox-support interface is a privileged interface designed for the
Canonical checkbox test and certification system. The system is able to run a
wide collection of system tests and is thus allowed to execute any command
mostly bypassing the sandbox.

The interface is allowed to run on both core and classic systems, so that
certification can use the same snap across the entire range of devices.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-34252

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>