osutil: add mount/libmount package with option validator
We need to recognize and validate libmount-style mount options, such as those
that one can use with the mount(1) utility on command line for possible
conflicts.  This is so that the apparmor parser, which does similar validation,
does not fail "late" in the interface setup process, when given incorrect data
by the mount-control interface.

Note that the Darwin implementation of the function is a stub.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-32984

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>