sandbox/apparmor: run apparmor via systemd-run
AppArmor parser is a compiler that transforms a tree of regular expressions
into one deterministic finite automaton. During the NFA to DFA transformation
phase a large amount of memory is required to compute the value and then to
minimize the resulting graph.

We are regularly hitting issues where, in continuous integration, snapd calls
apparmor_parser while being bound by a tight memory limit imposed on the
snapd.service unit. When the parser is asked to process a number of profiles it
will internally perform such computation in parallel. With an unlucky set,
snapd may get killed by the out-of-memory mechanism because the memory of the
parser is accounted to the same control group where snapd is running.

While this does not change the actual memory used, it does make the original
intent of catching runaway snapd memory usage, which may indicate a memory
leak, more useful.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-8640

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>