i/prompting: always validate inputs before handling simple cases
Previously, if a rule entry was expired, we'd discard it before
validation, to avoid the error which wasn't particularly relevant, since
the entry was already expired. And that validation is only called during
initial loading of rules. But if it's invalid, perhaps the expiration is
malformed as well. So validate first.

Similarly, if a client asks to patch a rule by removing a permission
which isn't valid, rather than do nothing, return an error that the
permission they requested to remove was invalid. Note that this is
different from the permission not existing in the rule: removing a valid
permission from a rule is idempotent, but asking to remove an invalid
permission (which snapd doesn't allow to exist in a rule) is an error.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>