o/ifacestate: record warning if prompting errors during startup
If an error occurs during interfaces requests manager initialization,
record a warning containing the error and indicating that prompting
will remain inactive until snapd is restarted.

Even when an error occurs, the `"apparmor-prompting"` feature flag is
left enabled, as the user deliberately turned this flag on, and we
don't want to change feature flags without user interaction.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>