tests: added nightly run from spread-cron (#14729)
Migrates all cron jobs from spread-cron to snapd that run spread tests on snapd. It makes a minor change in the spread-test.yaml to allow for use of the master snap along with two required environment variables.
It also adds the possibility of executing with (in addition to manually specifying the systems) "ALL" systems or "NONE." To help not create confusion, it will no longer be possible to leave the systems input blank.