i/builtin: check mount flags for conflicts
The mount-control interface used to allow requesting contradicting mount
options, such as nodev and dev and would silently pass this down to
apparmor_parser. The parser would detect this and cause the failure of the
security setup.

With this code we now detect this much earlier, at interface validation stage.

Fixes: https://warthogs.atlassian.net/browse/SNAPDENG-32984

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>